+++ api/src/main/java/com/qaengine/controllers/QuestionController.java
@@ -1,5 +1,7 @@
 package com.qaengine.controllers;
 
+import com.qaengine.Application;
+import com.qaengine.exceptions.BadRequestException;
 import com.qaengine.exceptions.ResourceNotFoundException;
 import com.qaengine.models.ApplicationUser;
 import com.qaengine.models.Category;
@@ -9,6 +11,7 @@ import com.qaengine.models.DTO.QuestionListDTO;
 import com.qaengine.services.CategoryService;
 import com.qaengine.services.QuestionService;
 import com.qaengine.services.UserService;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -65,7 +68,10 @@ public class QuestionController {
     }
 
     @DeleteMapping("/{id}")
-    public Long deleteQuestion(@PathVariable Long id) {
+    public Long deleteQuestion(@PathVariable Long id, Principal principal) {
+        if (!questionService.getQuestion(id).getUser().getUsername().equals(principal.getName())) {
+            throw new BadRequestException("Permission denied");
+        }
         try {
             questionService.deleteQuestion(id);
             return id;
@@ -77,8 +83,12 @@ public class QuestionController {
     @PutMapping("/{id}")
     public Question updateQuestion(
             @PathVariable Long id,
-            @RequestBody @Valid QuestionDTO questionInput
+            @RequestBody @Valid QuestionDTO questionInput,
+            Principal principal
     ) {
+        if (!questionService.getQuestion(id).getUser().getUsername().equals(principal.getName())) {
+            throw new BadRequestException("Permission denied");
+        }
         return questionService.updateQuestion(id, questionInput);
     }
